+++ example/example.php
@@ -2,8 +2,6 @@
 require_once __DIR__ . '/../vendor/autoload.php';
 use SevenEcks\Tableify\Tableify;
 
-$tableify = new Tableify();
-
 $data = [
     ['Name', 'Date', 'Phone', 'Age'], 
     ['Altec Lansing', '03/22/18', '617-555-0584', '30'],
@@ -17,22 +15,16 @@ $data = [
     ['Juicy Vee', '03/22/18', '978-555-0584', '34'],
 ];
 
-echo "Table Construction using default values on class:\n";
-$table_rows = $tableify->make($data);
-foreach ($table_rows as $row) {
+echo "Table Construction using default values on class and no method chaining:\n";
+$table = Tableify::new($data);
+$table = $table->make();
+$table_data = $table->toArray();
+foreach ($table_data as $row) {
     echo $row . "\n";
 }
 
 echo "Table Construction using method chaining:\n";
-$table_rows = $tableify->center()->seperatorPadding(2)->seperator('*')->headerCharacter('@')->make($data);
+$table_rows = Tableify::new($data)->center()->seperatorPadding(2)->seperator('*')->headerCharacter('@')->make()->toArray();
 foreach ($table_rows as $row) {
     echo $row . "\n";
 }
-
-// manual construction
-echo "Manual Table Construction using ->tableify:\n";
-$table_rows = $tableify->tableify($data, 'left', 1, '|', '-', '-');
-foreach ($table_rows as $row) {
-    echo $row . "\n";
-}
-