+++ src/Tableify.php
@@ -5,11 +5,11 @@ use SevenEcks\StringUtils\StringUtils;
 
 /**
  * This class allows you to create pretty printable / loggable tables from
- * a multi-dimensional array of strings!
+ * a multi-dimensional array of strings using method chaining!
  *
  * @author Brendan Butts <bbutts@stormcode.net>
  */
-class Tableify
+class Tableify implements TableifyInterface
 {
     // string library
     protected $su = null;
@@ -23,6 +23,10 @@ class Tableify
     protected $header_character = '-';
     // what should the row below the header row be made of
     protected $below_header_character = '-';
+    // the data for building the table
+    private $data = [];
+    // hold the created table data
+    private $table = [];
 
     /**
      * Construct the Tableify object, accepting a StringUtils class, 
@@ -31,12 +35,49 @@ class Tableify
      * @param object $string_utils;
      * @return void
      */
-    public function __construct($string_utils = null)
+    public function __construct(StringUtils $su = null)
     {
-        if (!$string_utils) {
-            $string_utils = new StringUtils();
+        // this is here for people who don't use :new to instantiate the
+        // class. It allows for dependency injection of $string_utils
+        if (!$su) {
+            $su = new StringUtils();
         }
-        $this->su = $string_utils;
+        // set the $string_utils package on the instance
+        $this->su = $su;
+    }
+
+    /**
+     * Static method for creating a new instance and passing in the $data,
+     * it also allows dependency injection for a $string_utils class
+     *
+     * @param array $data
+     * @param object $string_utils
+     * @return $this
+     */
+    public static function new (array $data, StringUtils $su = null)
+    {
+        // if no string utils exists, create one
+        if (!$su) {
+            $su = new StringUtils();
+        }
+        // create a new instance of this static class and provide the $string_utils
+        $instance = new static($su);
+        // set the $data on our new instance
+        $instance->setData($data);
+        // return the new instance object
+        return $instance;
+    }
+
+    /**
+     * Set the data array on the object
+     *
+     * @param array $data
+     * @return $this
+     */
+    public function setData(array $data)
+    {
+        $this->data = $data;
+        return $this;
     }
 
     /**
@@ -119,53 +160,35 @@ class Tableify
     }
 
     /**
-     * Use the method chained arguments (or defaults) to take the $data array
-     * that was passed into the function and turn it into a tableified array
-     * of rows of strings ready for pretty printing or logging
+     * Make a table based on the values assigned to the class via 
+     * methods.
      *
-     * @param array $data
-     * @return array
+     * @return $this
      */
-    public function make(array $data)
-    {
-        return $this->tableify($data, $this->formatter, $this->seperator_padding, $this->seperator, $this->header_character, $this->below_header_character);
-    }
-
-    /**
-     * Take a multidimensional array and turn it into an array of lines, using the first interior array as the header.
-     *
-     * @param array $data
-     * @param string $formatter
-     * @param int $seperator_padding
-     * @param string $seperator
-     * @param string $header_character
-     * @param string $below_header_character
-     * @return array
-     */
-    public function tableify(array $data, string $formatter = 'left', int $seperator_padding = 1, string $seperator = '|', string $header_character = '-',  string $below_header_character = '-')
+    public function make()
     {
         // empty table to start
         $table = [];
         // get the max length of each column in the table
-        $field_lengths = $this->maxMultidimensionalArrayStringSize($data);
+        $field_lengths = $this->maxMultidimensionalArrayStringSize($this->data);
         // get the length of a row by mocking up a row
-        $row_length = $this->calculateRowLength($field_lengths, $seperator_padding, $seperator);
+        $row_length = $this->calculateRowLength($field_lengths, $this->seperator_padding, $this->seperator);
         // get the top and bottom of the table
-        $book_end = $this->su->fill($row_length, $header_character);
+        $book_end = $this->su->fill($row_length, $this->header_character);
         // get the row that appears below the header row to seperate it from the other rows
-        $below_header = $this->su->fill($row_length, $below_header_character);
+        $below_header = $this->su->fill($row_length, $this->below_header_character);
         // add the top $book_end
         $table[] = $book_end;
         // loop over each 'row' in our array
-        for ($i = 0; $i < count($data); $i++) {
+        for ($i = 0; $i < count($this->data); $i++) {
             // line starts as just the left most seperator
-            $line = $seperator . $this->su->fill($seperator_padding);
+            $line = $this->seperator . $this->su->fill($this->seperator_padding);
             // loop over each 'field' aka column in our row
-            for ($j = 0; $j < count($data[$i]); $j++) {
+            for ($j = 0; $j < count($this->data[$i]); $j++) {
                 // format the fields such that they have a max length of the biggest field in this column
                 // and add the additional padding, then pass this to the formatter method and append it to
                 // the line and then add the seperator
-                $line .= $this->su->{$formatter}($data[$i][$j], $field_lengths[$j]) . $this->su->fill($seperator_padding) . $seperator . $this->su->fill($seperator_padding);
+                $line .= $this->su->{$this->formatter}($this->data[$i][$j], $field_lengths[$j]) . $this->su->fill($this->seperator_padding) . $this->seperator . $this->su->fill($this->seperator_padding);
             }
             // add the new line to the table
             $table[] = $line;
@@ -176,7 +199,19 @@ class Tableify
         }
         // add the bottom book end to the table
         $table[] = $book_end;
-        return $table;
+        // set the table data on the instance
+        $this->table = $table;
+        return $this;
+    }
+
+    /**
+     * Return the table data array for this instance
+     *
+     * @return array
+     */
+    public function toArray()
+    {
+        return $this->table;
     }
 
     /**