+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.4 - 2018-04-04
+- removed tablify method, only using ->make() from here on out
+- added toArray method
+- added TableifyInterface
+- added StringUtils typehint
+- added new() method for static instantation
+- refactored code
+- make() returns an object instead of an array
+- make stores table data in private $data prop
+- updated readme
+- updated example.php
+- removed images director
+
 ## 0.0.3 - 2018-04-04
 - Updated changelog
 - Updated readme to include API